refactor: extract expect utilities to separate module
Extracted expectUtils from bot.ts to improve code organization:
- expectBlock, expectBlockOfType
- expectItemInInventory, equipItem
- expectEntity, expectSufficientItems
- Helper functions: formatPosition, getInventorySummary, getEntityName, isBlockEmpty

These utilities provide consistent error handling for common validation patterns.
